requirements.txt
asgiref==3.6.0
autopep8==2.0.1
Django==3.2
django-cors-headers==3.13.0
djangorestframework>=3.12.4,<3.13
psycopg2>=2.8.6,<2.9
drf-nested-routers==0.93.4
flake8==6.0.0
mccabe==0.7.0
Pillow==9.4.0
pycodestyle==2.10.0
pyflakes==3.0.1
pytz==2022.7.1
sqlparse==0.4.3
tomli==2.0.1
tzdata==2022.7
drf-spectacular==0.25.1
